# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Fleet Management API",
    "version": "1.0.0",
    "description": "API documentation for the Fleet Management System"
  },
  "servers": [
    {
      "url": "http://localhost:3003",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    { "bearerAuth": [] }
  ],
  "paths": {
...
    "/api/v1/diesel-receiving/{id}": {
      "get": { "summary": "Get diesel receiving record by ID", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Diesel receiving record found" }, "404": { "description": "Not found" } } },
      "put": { "summary": "Update diesel receiving record by ID", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Diesel receiving record updated" }, "404": { "description": "Not found" } } },
      "delete": { "summary": "Delete diesel receiving record by ID", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "204": { "description": "Diesel receiving record deleted" }, "404": { "description": "Not found" } } }
    },
...
  }
}